move useEffect to searchPage,move loadData func
