+++ src/components/loader/styles.js
@@ -1,13 +1,12 @@
 import styled, { keyframes } from 'styled-components';
 
 const Rotate = keyframes`
-        0% {
-        transform: rotate(0);
-    }
-
-    100% {
-        transform: rotate(360deg);
-    }
+  0% {
+    transform:rotate(0);
+  }
+  100% {
+    transform:rotate(360deg);
+  }
 `;
 
 export default styled.div`
