+++ src/constants/config.js
@@ -1,4 +1,3 @@
 export default {
-  BASE_URL:
-    'https://cors-anywhere.herokuapp.com/https://api.pushshift.io/reddit/',
+  BASE_URL: 'https://api.pushshift.io/reddit/',
 };
