+++ src/pages/Search/components/Form.js
@@ -1,33 +1,18 @@
-import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import Button from '../../../components/button';
-import Loader from '../../../components/loader';
 import {
   FormContainer, InputGroup, Prefixed, Input,
 } from './Form.styles';
 
-const Form = ({ loadData }) => {
+const Form = () => {
   const { push } = useHistory();
   const { subreddit } = useParams();
   const [query, setQuery] = useState(subreddit);
-  const [loading, setLoading] = useState(false);
 
-  useEffect(() => {
-    const onLoad = async () => {
-      setLoading(true);
-      await loadData(query);
-      setLoading(false);
-    };
-    onLoad();
-  }, []);
-
-  const handleSubmit = async (e) => {
+  const handleSubmit = (e) => {
     e.preventDefault();
-    setLoading(true);
     push(`/search/${query}`);
-    await loadData(query);
-    setLoading(false);
   };
 
   const handleChange = ({ target }) => setQuery(target.value);
@@ -42,13 +27,8 @@ const Form = ({ loadData }) => {
           <Button>Search</Button>
         </InputGroup>
       </FormContainer>
-      {loading && <Loader />}
     </>
   );
-};
-
-Form.propTypes = {
-  loadData: PropTypes.func.isRequired,
 };
 
 export default Form;
